ref: fully switch collection management to LangChain
